Move helper classes into shared file and rename
The answer validator classes are also identical
between gitops and normal namespaces, so those
can go into the shared file too.

Now that it's not just questions, I've renamed
the shared file.
